opkg-build: create dpkg-deb compatible ipks/opks again
With a change from last august in dpkg-deb, the packages generated by opkg-build will not be accepted by dpkg-deb anymore.
The change disallows that the data.tar.gz is packed into the ipk/opk before the control.tar.gz.
For details have a look here:
http://anonscm.debian.org/gitweb/?p=dpkg/dpkg.git;a=commitdiff;h=547dca4c3ff23df5dfa554f1943b371cd7056ee4

For details which order is correct, see "man 5 deb", but in short it is:
debian-binary, control.tar.gz and data.tar.gz

Signed-off-by: Thilo Cestonaro <thilo@cestona.ro>
Signed-off-by: Paul Barker <paul@paulbarker.me.uk>