opkg-compare-indexes: Fix usage message
The usage message had the name of the program wrong.

If the number of arguments is not correct then the current code does not show
the usage message as indended. Instead it gives an index out of range error if
too few arguments are provided or silently ignores extra arguments. This code is
fixed so that the length of the arguments list is checked before arguments are
accessed, preventing errors and allowing the usage message to be shown
correctly.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>