opkg-show-deps: Add comments to explain code
There was some confusion discussed on the mailing list about what the various
loops in this script were doing. Comments are added to explain the behaviour,
hopefully preventing any confusion in the future. Thanks to Pavel Zubarev for
bringing this up and helping me to clarify what the current code is doing.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>