opkg-utils: Switch to python3
Switch all py script shebangs to `#!/usr/bin/env python3`.

Testing: Successfully ran the following on an IPK feed in $PWD:
 * opkg-make-index --checksum md5 --checksum sha256 -p Packages -m -l Packages.filelist .
 * opkg-graph-deps ./Packages
 * opkg-compare-indexes ./ ./
 * opkg-list-fields font-alias_1.0.3-r2.3.204_all.ipk
 * opkg-show-deps -p Packages font-alias

I'm not sure how to correctly run opkg-update-index and opkg-unbuild.
I attempted to run both with `-h`, no parse errors.

My motivation for this change is to add xz support to opkg filelists,
which requires python3's version of tarfile. See previous commits in
this series for details. However, this is probably a good change in
general since python2 support is starting to dwindle.

Signed-off-by: Haris Okanovic <haris.okanovic@ni.com>